The Go LIFX Client: Part 1
I need a LIFX LAN protocol client in Go that works as expected. The best-looking one has some quirks with the new protocol, so I've decided to experiment with the protocol by writing my own.

This is the first blog post documenting the writin of that client. It's also my first post in way too long.